Fixed a few conversion warnings

+++ src/realm/alloc_slab.cpp
@@ -40,7 +40,7 @@ SlabAlloc::SlabAlloc()
     size_t max = std::numeric_limits<size_t>::max();
     m_num_section_bases = 1 + get_section_index(max);
     m_section_bases.reset( new size_t[m_num_section_bases] );
-    for (int i = 0; i < m_num_section_bases; ++i) {
+    for (size_t i = 0; i < m_num_section_bases; ++i) {
         m_section_bases[i] = compute_section_base(i);
     }
 }
